requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.34
psycopg[binary]==3.2.10
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
redis==5.0.8
rq==1.16.2
httpx==0.27.2
stripe==10.5.0
cachetools==5.5.0
loguru==0.7.2

